feat(support api keys)!: implement API keys support
It is now possible to add drivers that require a key for the API that
they use.

BREAKING CHANGE: As it wasn't possible to keep tickers() backwards compatible,
since drivers that require a key throw an exception when it is missing, we
have decided to remove it, meaning the exported drivers should now be used
directly. Furthermore, as the change was already breaking, the lowercasing
for the driver names in list() has been removed, because this was only done
for backwards compatibility.